Worked on #734 (Virtualised (cached) grid loader) Refactoring Code from BOMapper to BOPropertyMapper and BORelationshipMapper so as to eliminate duplicate code. This is required so that I can implement the Loading of a Grid from the CachedBindingList while still using Habanero's Grid. Added Comments to UIGridColumn. Refactored Code from EditableDataSetProvider to use BOPropertyMapper so as to eliminate duplicate code.

+++ source/Habanero.Test.Bo/TestPack/ContactPersonTestBO.cs
@@ -894,6 +894,10 @@ namespace Habanero.Test.BO
             {
                 return;
             }
+            if(BORegistry.DataAccessor is DataAccessorMultiSource)
+            {
+                return;
+            }
             string sql = "DELETE FROM contact_person_address";
             DatabaseConnection.CurrentConnection.ExecuteRawSql(sql);
             sql = "DELETE FROM Contact_Person";
